requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
pydantic
uvicorn
sse_starlette
langchain-ollama
ipykernel
pypdf
chromadb
faiss-cpu
langgraph
langsmith
langchain-tavily